% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fwf_blaise.R
\name{read_fwf_blaise}
\alias{read_fwf_blaise}
\title{Read a fixed width ascii datafile using a blaise datamodel}
\usage{
read_fwf_blaise(datafile, modelfile, locale = readr::locale())
}
\arguments{
\item{datafile}{the ascii file with the data}

\item{modelfile}{the datamodel describing the data}

\item{locale}{locale as specified with readr::locale(). Uses "." as default
decimal separator. Can be used to change date_format, timezone, encoding, etc.}
}
\description{
Use this function to read a fwf that is described by a blaise datamodel.
Only handles files with a single datamodel.
If this function throws a warning, try using readr::problems() on the result,
this will usually show an error in the used locale.
}
\details{
Handles the following types:
\itemize{
  \item STRING
  \item INTEGER
  \item REAL
  \item DATETYPE
  \item ENUM (if numbered it will be converted to a factor with the numbers as labels)
  \item custom types (same as a numbered ENUM)
  }
}
\examples{
model = "
DATAMODEL Test
FIELDS
A     : STRING[1]
B     : INTEGER[1]
C     : REAL[3,1]
D     : REAL[3]
E     : (Male, Female)
F     : 1..20
G     : 1.00..100.00
ENDMODEL
"
data =
"A12,3,121 1  1,00
B23,41,2210 20,20
C34,512,120100,00"

blafile = tempfile('testbla', fileext = '.bla')
writeLines(model, con = blafile)
datafile = tempfile('testdata', fileext = '.asc')
writeLines(data, con = datafile)

df = read_fwf_blaise(datafile, blafile)
unlink(blafile)
unlink(datafile)

}
