% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_fwf_blaise.R
\name{write_fwf_blaise}
\alias{write_fwf_blaise}
\title{Write a fixed width ascii datafile and accompanying blaise datamodel}
\usage{
write_fwf_blaise(df, output_data, output_model = NULL,
  locale = readr::locale(decimal_mark = ","), force_string = FALSE)
}
\arguments{
\item{df}{dataframe to write}

\item{output_data}{path and name to output datafile. Will add .asc if no extension}

\item{output_model}{path and name to output datamodel. If NULL will use the
same name as output_data with .bla extension.}

\item{locale}{readr locale. Will use "," as a decimal mark by default, all
other defaults the same as in readr package}

\item{force_string}{If TRUE, will force unknown/unsupported column types to
string, otherwise these will throw an error}
}
\description{
Write a datafile in the blaise format (fwf ascii without separators)
will always write out a blaise datamodel describing the datafile as well
}
\examples{
datafilename = tempfile('testdata', fileext = '.asc')
blafilename = tempfile('testbla', fileext = '.bla')
data = data.frame(1, 1:10, sample(LETTERS[1:3], 10, replace = TRUE))
# would automatically create a bla file, but fails because factor is an
# unsupported type
write_fwf_blaise(data, datafilename)
# Will convert factor to string type
write_fwf_blaise(data,
  datafilename,
  blafilename,
  force_string = TRUE)

}
