% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_blaise_asc.R
\name{read_blaise_asc}
\alias{read_blaise_asc}
\title{Read a fixed width ascii datafile using a blaise datamodel}
\usage{
read_blaise_asc(datafile, modelfile, locale = readr::locale(decimal_mark =
  ","))
}
\arguments{
\item{datafile}{the ascii file with the data}

\item{modelfile}{the datamodel describing the data}

\item{locale}{locale as specified with readr::locale(). Uses "," as default
decimal separator. Can be used to change date_format, timezone, encoding, etc.}
}
\description{
Use this function to read a fwf that is described by a blaise datamodel.
Only handles files with a single datamodel.
}
\details{
Currently handles the following types:
\itemize{
  \item STRING
  \item INTEGER
  \item REAL
  \item DATETYPE
  }

Other types will throw an error
}
