% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_blaise_asc.R
\name{read_blaise_asc}
\alias{read_blaise_asc}
\title{Read a fixed width ascii datafile with a blaise datamodel}
\usage{
read_blaise_asc(datafile, modelfile)
}
\arguments{
\item{datafile}{the ascii file with the data}

\item{modelfile}{the datamodel describing the data}
}
\description{
Read a fixed width ascii datafile with a blaise datamodel
}
